+++ src/snovault/elasticsearch/create_mapping.py
@@ -13,7 +13,10 @@ from snovault import (
     TYPES,
 )
 from snovault.schema_utils import combine_schemas
-from .interfaces import ELASTIC_SEARCH
+from .interfaces import (
+    ELASTIC_SEARCH,
+    RESOURCES_INDEX,
+)
 import collections
 import json
 import logging
@@ -450,6 +453,10 @@ def set_index_mapping(es, index, doc_type, mapping):
     es.indices.put_mapping(index=index, doc_type=doc_type, body=mapping, ignore=[400], request_timeout=300)
 
 
+def create_snovault_index_alias(es, indices):
+    es.indices.put_alias(index=','.join(indices), name=RESOURCES_INDEX, request_timeout=300)
+
+
 def run(app, collections=None, dry_run=False):
     index = app.registry.settings['snovault.elasticsearch.index']
     registry = app.registry
@@ -460,6 +467,7 @@ def run(app, collections=None, dry_run=False):
     if not collections:
         collections = ['meta'] + list(registry[COLLECTIONS].by_item_type.keys())
 
+    indices = []
     for collection_name in collections:
         if collection_name == 'meta':
             doc_type = 'meta'
@@ -476,6 +484,10 @@ def run(app, collections=None, dry_run=False):
             continue
         create_elasticsearch_index(es, index, index_settings())
         set_index_mapping(es, index, doc_type, {doc_type: mapping})
+        if collection_name != 'meta':
+            indices.append(index)
+
+    create_snovault_index_alias(es, indices)
 
 
 def main():