+++ src/snovault/elasticsearch/indexer.py
@@ -19,16 +19,23 @@ from .interfaces import (
     ELASTIC_SEARCH,
     INDEXER
 )
+from .indexer_state import (
+    IndexerState,
+    all_uuids,
+    all_types,
+    SEARCH_MAX
+)
 import datetime
 import logging
 import pytz
 import time
 import copy
+import json
+import requests
 
 es_logger = logging.getLogger("elasticsearch")
 es_logger.setLevel(logging.ERROR)
 log = logging.getLogger(__name__)
-SEARCH_MAX = 99999  # OutOfMemoryError if too high
 MAX_CLAUSES_FOR_ES = 8192
 
 def includeme(config):
@@ -37,261 +44,72 @@ def includeme(config):
     registry = config.registry
     registry[INDEXER] = Indexer(registry)
 
+def get_related_uuids(request, es, updated, renamed):
+    '''Returns (set of uuids, False) or (list of all uuids, True) if full reindex triggered'''
 
-class IndexerState(object):
-    # Keeps track of uuids and indexer state by cycle.  Also handles handoff of uuids to followup indexer
-    def __init__(self, es, index, title='primary'):
-        self.es = es
-        self.index = index  # "index where indexerstate is stored"
+    updated_count = len(updated)
+    renamed_count = len(renamed)
+    if (updated_count + renamed_count) > MAX_CLAUSES_FOR_ES:
+        return (list(all_uuids(request.registry)), True)  # guaranteed unique
+    elif (updated_count + renamed_count) == 0:
+        return (set(), False)
 
-        self.title           = title
-        self.state_id        = self.title + '_indexer'       # State of the current or last cycle
-        self.todo_set        = self.title + '_in_progress'   # one cycle of uuids, sent to the Secondary Indexer
-        #self.failed_set      = self.title + '_failed'
-        #self.done_set        = self.title + '_done'          # Trying to get all uuids from 'todo' to this set
-        self.troubled_set    = self.title + '_troubled'      # uuids that failed to index in any cycle
-        self.last_set        = self.title + '_last_cycle'    # uuids in the most recent finished cycle
-        self.success_set     = None                          # None is the same as self.done_set
-        self.cleanup_this_cycle = [self.todo_set]  # ,self.failed_set,self.done_set]  # Clean up at end of current cycle
-        self.cleanup_last_cycle = [self.last_set,self.troubled_set]              # Clean up at beginning of next cycle
-        self.override           = 'reindex_' + self.title      # If exists then reindex all
-        # DO NOT INHERIT! These keys are for passing on to other indexers
-        self.followup_prep_list  = 'primary_followup_prep_list' # Setting up the uuids to be handled by a followup process
-        self.followup_ready_list = 'staged_by_primary_list'     # Followup list is added to here to pass baton
-        self.clock = {}
-        # some goals:
-        # 1) Hand-off to secondary: working
-        # 2) Detect and recover from interrupted cycle - working but noisy
-        #    Investigate es.count(xmin) to see if it is worth the restart=True flag
-        #    "The _version field is not indexed, therefore you cannot query/filter on it."
-        #    Consider adding xmin to body?
-        # 3) Identify what is being currently worked on - Too expensive beyond invalidated list
-        # 4) Record (double?) failures and consider blacklisting them - not tried, could do.
+    es.indices.refresh('_all')
 
-    # Private-ish primitives...
-    def get_obj(self, id):
-        try:
-            return self.es.get(index=self.index, doc_type='meta', id=id).get('_source',{})  # TODO: snovault/meta
-        except:
-            return {}
-
-    def put_obj(self, id, obj):
-        try:
-            self.es.index(index=self.index, doc_type='meta', id=id, body=obj)
-        except:
-            log.warn("Failed to save to es: " + id, exc_info=True)
-
-    def delete_objs(self, ids):
-        for id in ids:
-            try:
-                self.es.delete(index=self.index, doc_type='meta', id=id)
-            except:
-                pass
-
-    def get_list(self, id):
-        return self.get_obj(id).get('list',[])
-
-    def get_count(self, id):
-        return self.get_obj(id).get('count',0)
-
-    def put_list(self, id, a_list):
-        return self.put_obj(id, { 'list': a_list, 'count': len(a_list) })
-
-    #def get_diff(self,orig_id, subtract_ids):
-    #    result_set = set(self.get_list(orig_id))
+    # TODO: batching may allow us to drive a partial reindexing much greater than 99999
+    #BATCH_COUNT = 100  # NOTE: 100 random uuids returned > 99999 results!
+    #beg = 0
+    #end = BATCH_COUNT
+    #related_set = set()
+    #updated_list = list(updated)  # Must be lists
+    #renamed_list = list(renamed)
+    #while updated_count > beg or renamed_count > beg:
+    #    if updated_count > end or beg > 0:
+    #        log.error('Indexer looking for related uuids by BATCH[%d,%d]' % (beg, end))
     #
-    #    if len(result_set) > 0:
-    #        for id in subtract_ids:
-    #            subtract_list = self.get_list(id)
-    #            if len(subtract_list):
-    #                result_set = result_set.difference(set(subtract_list))
-    #    return result_set
-
-    #def set_add(self, id, vals):
-    #    set_to_update = set(self.get_list(id))
-    #    if len(set_to_update) > 0:
-    #        set_to_update.update(vals)
-    #    else:
-    #        set_to_update = set(vals)
-    #    self.put_list(id, set_to_update)
-
-    def list_extend(self, id, vals):
-        list_to_extend = self.get_list(id)
-        if len(list_to_extend) > 0:
-            list_to_extend.extend(vals)  # TODO: consider capping at SEARCH_MAX (keeping count but not uuids).  Requires followup handoff work.
-        else:
-            list_to_extend = vals
-
-        self.put_list(id, list_to_extend)
-
-    def rename_objs(self, from_id, to_id):
-        val = self.get_list(from_id)
-        if val:
-            self.put_list(to_id, val)
-            self.delete_objs([from_id])
-
-    # Public access...
-    def get(self):
-        '''Returns the basic state info'''
-        return self.get_obj(self.state_id)
-
-    def put(self, state):
-        '''Update the basic state info'''
-        # Don't save errors in es
-        errors = state.pop('errors', None)
-
-        state['title'] = self.state_id
-        self.put_obj(self.state_id, state)
-
-        if errors is not None:
-            state['errors'] = errors
-
-    def get_initial_state(self):
-        '''Useful to initialize at idle cycle'''
-        new_state = { 'title': self.state_id, 'status': 'idle'}
-        state = self.get()
-        for var in ['cycles']:  # could expand this list
-            val = state.pop('cycles',None)
-            if val is not None:
-                new_state[var] = val
-        return new_state
-
-    def start_clock(self, name):
-        '''Can start a named clock and use it later to figure out elapsed time'''
-        self.clock[name] = datetime.datetime.now(pytz.utc)
-
-    def elapsed(self, name):
-        '''Returns string of time elapsed since named clock started.'''
-        start = self.clock.get(name)
-        if start is None:
-            return 'unknown'
-        else:
-            return str(datetime.datetime.now(pytz.utc) - start)
-
-    def priority_cycle(self, registry):
-        '''Initial startup, override, or interupted prior cycle can all lead to a priority cycle.
-           returns (discovered xmin, uuids, whether previous cycle was interupted).'''
-        # Not yet started?
-        initialized = self.get_obj("indexing")
-        if not initialized:
-            self.delete_objs([self.override, self.followup_ready_list])
-            state = self.get()
-            state['status'] = 'uninitialized'
-            self.put(state)
-            return (-1, [], False)  # primary indexer will know what to do and secondary indexer should do nothing yet
-
-        state = self.get()
-
-        # Rare call for indexing all...
-        override = self.get_obj(self.override)
-        if override:
-            self.delete_objs([self.override,self.followup_ready_list])
-            uuids = list(all_uuids(registry))
-            log.warn('%s override doing all: %d' % (self.state_id, len(uuids)))
-            return (-1, uuids, False)
-
-        if state.get('status', '') != 'indexing':
-            return (-1, [], False)
-
-        xmin = state.get('xmin', -1)
-        #snapshot = state.get('snapshot', None)
-        if xmin == -1:  # or snapshot is None:
-            return (-1, [], False)
-
-        #assert(self.get_count(self.done_set) == 0)  # Valid for cycle-level accounting only
-        #undone_uuids = self.get_diff(self.todo_set, [self.done_set])  # works for any accountingu
-        undone_uuids = self.get_list(self.todo_set)                    # works fastest for cycle-level accounting
-        if len(undone_uuids) <= 0:  # TODO SEARCH_MAX?  SEARCH_MAX/10
-            return (-1, [], False)
-
-        # Note: do not clean up last cycle yet because we could be restarted multiple times.
-        return (xmin, undone_uuids, True)
+    #    updated = []
+    #    if updated_count > beg:
+    #        updated = updated_list[beg:end]
+    #    renamed = []
+    #    if renamed_count > beg:
+    #        renamed = renamed_list[beg:end]
+    #
+    #     search ...
+    #     accumulate...
+    #
+    #    beg += BATCH_COUNT
+    #    end += BATCH_COUNT
 
 
-    def prep_for_followup(self, xmin, uuids):
-        '''Prepare a cycle of uuids for passing to a followup indexer (e.g. audits, viscache)'''
-        prep_list = [ "xmin:%s" % xmin ]
-        prep_list.extend(uuids)
-        self.put_list(self.followup_prep_list, prep_list)
-        # No need to preserve anything on the prep_list as it passes to the staged list in one cycle.
+    res = es.search(index='_all', size=SEARCH_MAX, request_timeout=60, body={
+        'query': {
+            'bool': {
+                'should': [
+                    {
+                        'terms': {
+                            'embedded_uuids': updated,
+                            '_cache': False,
+                        },
+                    },
+                    {
+                        'terms': {
+                            'linked_uuids': renamed,
+                            '_cache': False,
+                        },
+                    },
+                ],
+            },
+        },
+        '_source': False,
+    })
 
-    def start_cycle(self, uuids, state=None):
-        '''Every indexing cycle must be properly opened.'''
-        self.clock = {}
-        self.start_clock('cycle')
-        if state is None:
-            state = self.get()
-        state['status'] = 'indexing'
-        state['cycle_count'] = len(uuids)
+    if res['hits']['total'] > SEARCH_MAX:
+        return (list(all_uuids(request.registry)), True)  # guaranteed unique
 
-        self.put(state)
-        self.delete_objs(self.cleanup_last_cycle)
-        self.delete_objs(self.cleanup_this_cycle)
-        self.put_list(self.todo_set, set(uuids))
-        return state
+    related_set = {hit['_id'] for hit in res['hits']['hits']}
 
-    def start_pass2(self, state):
-        state['pass1_took'] = self.elapsed('cycle')
-        self.put(state)
-        self.start_clock('pass2')
-        return state
+    return (related_set, False)
 
-
-    def add_errors(self, errors, finished=True):
-        '''To avoid 16 worker concurency issues, errors are recorded at the end of a cycle.'''
-        uuids = [err['uuid'] for err in errors]  # better be uuids!
-        if len(uuids) > 0:
-            # Forget extend, this is cycle-level accounting
-            #self.list_extend(self.failed_set, uuids)
-            # Forget failed_set, this is cycle-level accounting
-            #self.put_list(self, self.failed_set, uuids)
-
-            if finished:
-                # Forget sets... cycle-level accounting so errors => failed_set => troubled_set all in one cycle.
-                # handle troubled uuids:
-                #troubled_uuids = set(self.get_list(self.failed_set))
-                #if len(troubled_uuids):
-                #    self.set_add(self.troubled_set, troubled_uuids)
-                #    # TODO: could make doubled_troubled set and use it to blacklist uuids
-                self.put_list(self.troubled_set, uuids)
-
-    def finish_cycle(self, state, errors=None):
-        '''Every indexing cycle must be properly closed.'''
-
-        if 'pass2' in self.clock.keys():
-            state['pass2_took'] = self.elapsed('pass2')
-
-        if errors:  # By handling here, we avoid overhead and concurrency issues of uuid-level accounting
-            self.add_errors(errors)
-
-        # pass any staged items to followup
-        if self.followup_prep_list is not None:
-            # TODO: send signal for 'all' when appropriate.  Saves the following expensive lines.
-            hand_off_list = self.get_list(self.followup_prep_list)
-            if len(hand_off_list) > 0:  # Have to push because ready_list may still have previous cycles in it
-                self.list_extend(self.followup_ready_list, hand_off_list)
-                self.delete_objs([self.followup_prep_list])
-
-        # cycle-level accounting so todo => done => last in this function
-        #self.rename_objs(self.todo_set, self.done_set)
-        done_count = self.get_count(self.todo_set)
-        self.rename_objs(self.todo_set, self.last_set)
-
-        if self.success_set is not None:
-            state[self.title + '_updated'] = self.get_count(self.success_set)
-        cycle_count = state.pop('cycle_count', None)
-        #assert(cycle_count == done_count)
-        state['indexed'] = done_count
-
-        #self.rename_objs(self.done_set, self.last_set)   # cycle-level accounting so todo => done => last in this function
-        self.delete_objs(self.cleanup_this_cycle)
-        state['status'] = 'done'
-        state['cycles'] = state.get('cycles', 0) + 1
-        state['cycle_took'] = self.elapsed('cycle')
-
-        self.put(state)
-
-        return state
 
 
 @view_config(route_name='index', request_method='POST', permission="index")
@@ -304,7 +122,6 @@ def index(request):
     recovery = request.json.get('recovery', False)
     es = request.registry[ELASTIC_SEARCH]
     indexer = request.registry[INDEXER]
-    stage_for_followup = request.registry.settings.get("stage_for_followup", False)
     session = request.registry[DBSESSION]()
     connection = session.connection()
     first_txn = None
@@ -313,19 +130,22 @@ def index(request):
     invalidated = []
     xmin = -1
 
-    # May have undone uuids from prior cycle
-    state = IndexerState(es,INDEX)
+    # Currently 2 possible followup indexers (base.ini [set stage_for_followup = vis_indexer, region_indexer])
+    stage_for_followup = list(request.registry.settings.get("stage_for_followup", '').replace(' ','').split(','))
 
+    # May have undone uuids from prior cycle
+    state = IndexerState(es, INDEX, followups=stage_for_followup)
+
+    (xmin, invalidated, restart) = state.priority_cycle(request)
     # OPTIONAL: restart support
-    #(xmin, invalidated, restart) = state.priority_cycle(request.registry)
+    if restart:  # Currently not bothering with restart!!!
+        xmin = -1
+        invalidated = []
     # OPTIONAL: restart support
 
     result = state.get_initial_state()  # get after checking priority!
-    if len(invalidated) > 0:
-        if xmin == -1:
-            xmin = get_current_xmin(request)
-        flush = True
-    else:
+
+    if xmin == -1 or len(invalidated) == 0:
         xmin = get_current_xmin(request)
 
         last_xmin = None
@@ -333,13 +153,22 @@ def index(request):
             last_xmin = request.json['last_xmin']
         else:
             status = es.get(index=INDEX, doc_type='meta', id='indexing', ignore=[400, 404])
-            if status['found']:
+            if status['found'] and 'xmin' in status['_source']:
                 last_xmin = status['_source']['xmin']
+        if last_xmin is None:  # still!
+            if 'last_xmin' in result:
+                last_xmin = result['last_xmin']
+            elif 'xmin' in result and result['xmin'] < xmin:
+                last_xmin = result['state']
 
         result.update(
             xmin=xmin,
             last_xmin=last_xmin,
         )
+
+    if len(invalidated) > SEARCH_MAX:  # Priority cycle already set up
+        flush = True
+    else:
 
         flush = False
         if last_xmin is None:
@@ -351,7 +180,7 @@ def index(request):
                 TransactionRecord.xid >= last_xmin,
             )
 
-            invalidated = set()
+            invalidated = set(invalidated)  # not empty if API index request occurred
             updated = set()
             renamed = set()
             max_xid = 0
@@ -366,52 +195,30 @@ def index(request):
                 renamed.update(txn.data.get('renamed', ()))
                 updated.update(txn.data.get('updated', ()))
 
+            if invalidated:        # reindex requested, treat like updated
+                updated |= invalidated
+
             result['txn_count'] = txn_count
-            if txn_count == 0:
+            if txn_count == 0 and len(invalidated) == 0:
+                state.send_notices()
                 return result
 
-            if len(updated) + len(renamed) > MAX_CLAUSES_FOR_ES:
-                invalidated = list(all_uuids(request.registry))
+            (related_set, full_reindex) = get_related_uuids(request, es, updated, renamed)
+            if full_reindex:
+                invalidated = related_set
                 flush = True
             else:
-                es.indices.refresh('_all')
-                res = es.search(index='_all', size=SEARCH_MAX, request_timeout=60, body={
-                    'query': {
-                        'bool': {
-                            'should': [
-                                {
-                                    'terms': {
-                                        'embedded_uuids': updated,
-                                        '_cache': False,
-                                    },
-                                },
-                                {
-                                    'terms': {
-                                        'linked_uuids': renamed,
-                                        '_cache': False,
-                                    },
-                                },
-                            ],
-                        },
-                    },
-                    '_source': False,
-                })
-
-                if res['hits']['total'] > SEARCH_MAX:
-                    invalidated = list(all_uuids(request.registry))
-                    flush = True
-                else:
-                    referencing = {hit['_id'] for hit in res['hits']['hits']}
-                    invalidated = referencing | updated
-                    result.update(
-                        max_xid=max_xid,
-                        renamed=renamed,
-                        updated=updated,
-                        referencing=len(referencing),
-                        invalidated=len(invalidated),
-                        txn_count=txn_count,
-                        first_txn_timestamp=first_txn.isoformat(),
-                    )
+                invalidated = related_set | updated
+                result.update(
+                    max_xid=max_xid,
+                    renamed=renamed,
+                    updated=updated,
+                    referencing=len(related_set),
+                    invalidated=len(invalidated),
+                    txn_count=txn_count
+                )
+                if first_txn is not None:
+                    result['first_txn_timestamp'] = first_txn.isoformat()
 
             if invalidated and not dry_run:
                 # Exporting a snapshot mints a new xid, so only do so when required.
@@ -421,7 +228,7 @@ def index(request):
                     snapshot_id = connection.execute('SELECT pg_export_snapshot();').scalar()
 
     if invalidated and not dry_run:
-        if stage_for_followup:
+        if len(stage_for_followup) > 0:
             # Note: undones should be added before, because those uuids will (hopefully) be indexed in this cycle
             state.prep_for_followup(xmin, invalidated)
 
@@ -429,18 +236,7 @@ def index(request):
 
         # Do the work...
 
-        ### OPTIONAL: audit via 2-pass is coming...
-        #log.info("indexer starting pass 1 on %d uuids", len(invalidated))
-        ### OPTIONAL: audit via 2-pass is coming...
         errors = indexer.update_objects(request, invalidated, xmin, snapshot_id, restart)
-
-        ### OPTIONAL: audit via 2-pass is coming...
-        #result = state.start_pass2(result)
-        #log.info("indexer starting pass 2 on %d uuids", len(invalidated))
-        #audit_errors = indexer.update_audits(request, invalidated, xmin, snapshot_id)  # ignore restart
-        #if len(audit_errors):
-        #    errors.extend(audit_errors)
-        ### OPTIONAL: audit via 2-pass is coming...
 
         result = state.finish_cycle(result,errors)
 
@@ -471,6 +267,7 @@ def index(request):
     if first_txn is not None:
         result['txn_lag'] = str(datetime.datetime.now(pytz.utc) - first_txn)
 
+    state.send_notices()
     return result
 
 
@@ -494,27 +291,6 @@ def get_current_xmin(request):
     # which is not available in recovery.
     xmin = query.scalar()  # lowest xid that is still in progress
     return xmin
-
-def all_uuids(registry, types=None):
-    # First index user and access_key so people can log in
-    collections = registry[COLLECTIONS]
-    initial = ['user', 'access_key']
-    for collection_name in initial:
-        collection = collections.by_item_type.get(collection_name, [])
-        # for snovault test application, there are no users or keys
-        if types is not None and collection_name not in types:
-            continue
-        for uuid in collection:
-            yield str(uuid)
-    for collection_name in sorted(collections.by_item_type):
-        if collection_name in initial:
-            continue
-        if types is not None and collection_name not in types:
-            continue
-        collection = collections.by_item_type[collection_name]
-        for uuid in collection:
-            yield str(uuid)
-
 
 class Indexer(object):
     def __init__(self, registry):
@@ -550,9 +326,6 @@ class Indexer(object):
 
         last_exc = None
         try:
-            ### OPTIONAL: audit via 2-pass is coming...
-            #doc = request.embed('/%s/@@index-data/noaudit/' % uuid, as_user='INDEXER')
-            ### OPTIONAL: audit via 2-pass is coming...
             doc = request.embed('/%s/@@index-data/' % uuid, as_user='INDEXER')
         except StatementError:
             # Can't reconnect until invalid transaction is rolled back
@@ -562,18 +335,6 @@ class Indexer(object):
             last_exc = repr(e)
 
         if last_exc is None:
-            ### OPTIONAL: audit via 2-pass is coming...
-            #try:
-            #    audit = self.es.get(index=self.index, id=str(uuid)).get('_source',{}).get('audit')  # Any version
-            #    if audit:
-            #        doc.update(
-            #            audit=audit,
-            #            audit_stale=True,
-            #        )
-            #except:
-            #    pass
-            ### OPTIONAL: audit via 2-pass is coming...
-
             for backoff in [0, 10, 20, 40, 80]:
                 time.sleep(backoff)
                 try:
@@ -601,94 +362,6 @@ class Indexer(object):
 
         timestamp = datetime.datetime.now().isoformat()
         return {'error_message': last_exc, 'timestamp': timestamp, 'uuid': str(uuid)}
-
-    def update_audits(self, request, uuids, xmin, snapshot_id=None):
-        # We need to make the recently pushed objects searchable
-        # Audit will fetch the most recent version after the refresh
-        self.es.indices.refresh(index='_all')
-        errors = []
-        for i, uuid in enumerate(uuids):
-            error = self.update_audit(request, uuid, xmin)
-            if error is not None:
-                errors.append(error)
-            if (i + 1) % 50 == 0:
-                log.info('Auditing %d', i + 1)
-
-        return errors
-
-    def update_audit(self, request, uuid, xmin):
-        request.datastore = 'elasticsearch'  # Audits are built on elastic search !!!
-
-        last_exc = None
-        # First get the object currently in es
-        try:
-            # This slings the single result via PickStorage via CachedModel via ElasticSearchStorage
-            # Only works if request.datastore = 'elasticsearch'
-            result = self.esstorage.get_by_uuid(uuid)
-            doc = result.source
-        except StatementError:
-            # Can't reconnect until invalid transaction is rolled back
-            raise
-        except Exception as e:
-            log.error("Error can't find %s in %s", uuid, ELASTIC_SEARCH)
-            last_exc = repr(e)
-
-        ### # Handle audits:
-        index_audit = False
-        if last_exc is None:
-            # It might be possible to assert that the audit is either empty or stale
-            # TODO assert('audit_stale' is in doc or doc.get('audit') is None)
-
-            try:
-                # using audit-now bypasses cached_views 'audit' ensureing the audit is recacluated.
-                audit = request.embed(str(uuid), '@@audit-now', as_user='INDEXER')['audit']
-                if audit or doc.get('audit_stale',False):
-                    doc.update(
-                        audit=audit,
-                        audit_stale=False,
-                    )
-                    index_audit = True
-            except StatementError:
-                print('statement error')
-                # Can't reconnect until invalid transaction is rolled back
-                raise
-            except Exception as e:
-                log.error('Error rendering /%s/@@index-audits', uuid, exc_info=True)
-                last_exc = repr(e)
-
-        if index_audit:
-            if last_exc is None:
-                for backoff in [0, 10, 20, 40, 80]:
-                    time.sleep(backoff)
-                    try:
-                        self.es.index(
-                            index=doc['item_type'], doc_type=doc['item_type'], body=doc,
-                            id=str(uuid), version=xmin, version_type='external_gte',
-                            request_timeout=30,
-                        )
-                    except StatementError:
-                        # Can't reconnect until invalid transaction is rolled back
-                        raise
-                    except ConflictError:
-                        #log.warning('Conflict indexing %s at version %d', uuid, xmin)
-                        # This case occurs when the primary indexer is cycles ahead of the secondary indexer
-                        # And this uuid will be secondarily indexed again on a second round
-                        # So no error, pretend it has indexed and move on.
-                        return  # No use doing any further secondary indexing
-                    except (ConnectionError, ReadTimeoutError, TransportError) as e:
-                        log.warning('Retryable error indexing %s: %r', uuid, e)
-                        last_exc = repr(e)
-                    except Exception as e:
-                        log.error('Error indexing %s', uuid, exc_info=True)
-                        last_exc = repr(e)
-                        break
-                    else:
-                        # Get here on success and outside of try
-                        return
-
-        if last_exc is not None:
-            timestamp = datetime.datetime.now().isoformat()
-            return {'error_message': last_exc, 'timestamp': timestamp, 'uuid': str(uuid)}
 
     def shutdown(self):
         pass